### 2025-09-05 #8929 @ivborissov
@crvernon Thanks for pointing out the word count issue. I’ve reduced the text and pushed the update to the master branch.
