### 2025-09-16 #8929 @ivborissov
Hi @crvernon I noticed that our paper was assigned to JOSS Track 5 (Data Science, Artificial Intelligence, and Machine Learning). However, the software package we are proposing focuses on uncertainty quantification in mathematical modeling, which I believe aligns more closely with Track 7 (Computer Science, Information Science, and Mathematics). Would it be possible to reassign the paper to Track 7?
