+++ test/test_interface.jl
@@ -1,24 +1,151 @@
-using LikelihoodProfiler, Test, Optimization
+using Test
+using LikelihoodProfiler
+using Optimization
 
-# ParameterProfile tests
-optprob1 = OptimizationProblem((x,p)->x[1]^2+x[2]^2, [1., 2.0])
+# ----------------------------
+# ParameterTarget constructors
+# ----------------------------
+@testset "ParameterTarget constructors" begin
+    @test_throws DimensionMismatch ParameterTarget(; idxs=1:2,
+        profile_lower=[-10], profile_upper=[10,10])
 
-@test_throws DimensionMismatch ProfileLikelihoodProblem(optprob1, [0.,0, 0.]) 
-@test_throws ArgumentError ProfileLikelihoodProblem(optprob1, [0.,0],  [(-5,4), (-2,1)]; threshold = -1)
-@test_throws DimensionMismatch ProfileLikelihoodProblem(optprob1, [0.,0], [(-5,4), (-2,1), (-3, 20)])
+    @test_throws ArgumentError ParameterTarget(; idxs=[1],
+        profile_lower=[7], profile_upper=[4])
 
-@test_throws ArgumentError solve(ProfileLikelihoodProblem(optprob1, [0.,0]), OptimizationProfiler(optimizer=Optimization.LBFGS(), stepper = FixedStep(; initial_step=0.01))) 
-@test_throws ArgumentError solve(ProfileLikelihoodProblem(optprob1, [0.,0], [(-Inf,4), (-2,1)]), OptimizationProfiler(optimizer=Optimization.LBFGS(), stepper = FixedStep(; initial_step=0.01)), idxs=[1])
+    @test_throws ArgumentError ParameterTarget(; idxs=[1,1],
+        profile_lower=[-1, -1], profile_upper=[1, 1])
 
-plprob = ProfileLikelihoodProblem(optprob1, [0.,0], (-5,4))
-plprob2 = remake(plprob, profile_range = [nothing, (-5,4)] )
-@test_throws ArgumentError solve(plprob2, OptimizationProfiler(optimizer=Optimization.LBFGS(), stepper = FixedStep(; initial_step=0.01)), idxs=[1])
+    @test_throws ArgumentError ParameterTarget(; idxs=Int[],
+        profile_lower=Float64[], profile_upper=Float64[])
 
-# FunctionProfile tests
-#=
-optf1 = OptimizationFunction((x,p)->x[1]+x[2])
-funprob = ProfileLikelihoodProblem(optprob1, [0.,0], optf1, (-2,1))
+    @test_throws ArgumentError ParameterTarget(; idxs=[0],
+        profile_lower=[-1], profile_upper=[1])
 
-@test_throws ArgumentError ProfileLikelihoodProblem(optprob1, [0.,0], optf1, (5,4)) 
-@test_throws ArgumentError remake(funprob, optpars=zeros(4))
-=#
+    # positive construction with vector bounds 
+    pt = ParameterTarget(; idxs=[1,3], profile_lower=[-2.0, -2.0], profile_upper=[2.0, 2.0])
+    @test pt.profile_lower == [-2.0, -2.0]
+    @test pt.profile_upper == [ 2.0,  2.0]
+end
+
+# ----------------------------
+# FunctionTarget constructors
+# ----------------------------
+@testset "FunctionTarget constructors" begin
+    f1 = OptimizationFunction((x,p)->x[1])
+    f2 = OptimizationFunction((x,p)->x[2])
+
+    # positive construction with vector bounds
+    ft = FunctionTarget(; fs=[f1,f2], profile_lower=[-1.0, -1.0], profile_upper=[1.0, 1.0])
[…]
