+++ src/problem_interface.jl
@@ -1,116 +1,305 @@
+"""
+   AbstractProfileLikelihoodProblem
 
-abstract type AbstractProfileLikelihoodProblem{T,PF} end
+Abstract type for all profile likelihood problems. Concrete subtypes carry the
+information required to define a specific profile likelihood problem.
+"""
+abstract type AbstractProfileLikelihoodProblem end
 
 """
-    ProfileLikelihoodProblem{T,probType,P,PF,PR}
+    AbstractProfileTarget
+
+Abstract type for all profile targets. Concrete subtypes carry the
+information required to define what is being profiled. 
+"""
+abstract type AbstractProfileTarget end
+
+"""
+    ParameterTarget{I,B}
+
+Profile target representing profiling of model parameters.
+
+### Fields
+
+- `idxs::AbstractVector{<:Integer}`: Indices of the parameters being profiled.
+- `profile_lower::AbstractVector{<:Real}`: Lower bounds for the profile likelihood. 
+- `profile_upper::AbstractVector{<:Real}`: Upper bounds for the profile likelihood. 
+
+Profile bounds `profile_lower` and `profile_upper` should be vectors of finite numerical values. 
+
+### Constructors
+
+Create a target with explicit lower and upper bounds for each index.
+```julia
+ParameterTarget(; idxs::AbstractVector{<:Integer}, profile_lower::AbstractVector{<:Real}, profile_upper::AbstractVector{<:Real})
+```
+"""
+struct ParameterTarget{I<:AbstractVector{<:Integer}, B<:AbstractVector{<:Real}} <: AbstractProfileTarget
+  idxs::I
+  profile_lower::B
+  profile_upper::B
+
+  function ParameterTarget(idxs::I, 
+                           profile_lower::AbstractVector{<:Real}, 
+                           profile_upper::AbstractVector{<:Real}) where {I<:AbstractVector{<:Integer}}
+    n = length(idxs)
+    n > 0           || throw(ArgumentError("`idxs` must be non-empty."))
+    allunique(idxs) || throw(ArgumentError("`idxs` must be unique."))
+    all(>=(1), idxs) || throw(ArgumentError("`idxs` must be positive integers."))
+    
+    lb_typed, ub_typed = _promote_profile_bounds(profile_lower, profile_upper, n)
+    new{I,typeof(lb_typed)}(idxs, lb_typed, ub_typed)
+  end
+end
+
+function Base.show(io::IO, mime::MIME"text/plain", target::ParameterTarget)
+  println(io, "ParameterTarget: $(length(get_profile_idxs(target))) parameter(s) to profile.")
+end
+
+"""
+    FunctionTarget{F,B}
+
+Profile target representing profiling of functions of model parameters.
+
+### Fields
+
+- `fs::AbstractVector{<:OptimizationFunction}`: Functions of the parameters being profiled.
+- `profile_lower::AbstractVector{<:Real}`: Lower bounds for the profile likelihood. 
+- `profile_upper::AbstractVector{<:Real}`: Upper bounds for the profile likelihood. 
+
+Profile bounds `profile_lower` and `profile_upper` should be vectors of finite numerical values. 
+
+### Constructors
+
+Create a target with explicit lower and upper bounds for each function of parameters.
+```julia
+FunctionTarget(; fs::AbstractVector{<:OptimizationFunction}, profile_lower::AbstractVector{<:Real}, profile_upper::AbstractVector{<:Real})
+```
+"""
+struct FunctionTarget{F<:AbstractVector{<:OptimizationFunction}, B<:AbstractVector{<:Real}} <: AbstractProfileTarget
+  fs::F
+  profile_lower::B
+  profile_upper::B
+
[…]
