+++ src/LikelihoodProfiler.jl
@@ -23,7 +23,7 @@ include("problem_interface.jl")
 include("profile_solution.jl")
 include("profile_methods.jl")
 include("profiler_state.jl")
-include("profile.jl")
+include("solve.jl")
 include("utils.jl")
 include("optprob_utils.jl")
 include("odeprob_utils.jl")
