### 2025-09-26 e2ea45e
new interface - tests pass
