+++ src/profiler_step.jl
@@ -1,4 +1,72 @@
-abstract type AbstractProfilerStep{S} end
+
+################################################## OPTIMIZATION PROFILER STEP ##################################################
+
+function profiler_step!(profiler_cache::ProfilerCache, solver_cache::OptimizationSolverCache)
+  @unpack opt_cache, stepper = solver_cache
+  idx = get_profile_idx(profiler_cache)
+
+  pars_guess = propose_next_pars!(profiler_cache, stepper)
+
+  fill_x_reduced!(opt_cache.reinit_cache.u0, pars_guess, idx)
+  set_x_fixed!(opt_cache.reinit_cache.p, pars_guess[idx])
+  
+  sol = solve_opt_cache(opt_cache)
+
+  if SciMLBase.successful_retcode(sol.retcode)
+    fill_x_full!(profiler_cache.pars_cur, sol.u, idx, pars_guess[idx])
+    profiler_cache.x_cur = pars_guess[idx]
+    profiler_cache.obj_cur = sol.objective
+    profiler_cache.iter += 1
+  else
+    @warn "Solver returned $(sol.retcode) retcode at profile point x = $(profiler_cache.x_cur). Profiling is interrupted."
+  end
+  solver_cache.retcode_original = sol.retcode
+  return nothing
+end
+
+function solve_opt_cache(opt_cache::OptimizationCache)  
+  if isempty(opt_cache.reinit_cache.u0)
+    return solve_empty_opt_cache(opt_cache)
+  else
+    return solve!(opt_cache)
+  end
+end
+
+function solve_empty_opt_cache(opt_cache::OptimizationCache)
+  u = opt_cache.reinit_cache.u0
+  p = opt_cache.reinit_cache.p
+  
+  t = @elapsed obj = opt_cache.f(u, p)
+  stats = SciMLBase.OptimizationStats(; iterations = 1, time = t, fevals = 1)
+
+  return SciMLBase.build_solution(opt_cache, opt_cache.opt, u, obj; 
+    retcode = ReturnCode.Success, stats)
+end
+
+################################################## INTEGRATION PROFILER STEP ##################################################
+
+function profiler_step!(profiler_cache::ProfilerCache, solver_cache::IntegrationSolverCache)
+  @unpack ode_cache, opt_cache = solver_cache
+
+  integrator = ode_cache
+  SciMLBase.step!(integrator)
+
+  if SciMLBase.successful_retcode(integrator.sol.retcode)
+    idx = get_profile_idx(profiler_cache)
+
+    profiler_cache.pars_cur .= view(integrator.u, 1:length(integrator.u)-1)
+    profiler_cache.x_cur = integrator.u[idx]
+    profiler_cache.obj_cur = evaluate_obj(get_plprob(profiler_cache), profiler_cache.pars_cur)
+    profiler_cache.iter += 1
+  else
+    @warn "Solver returned $(integrator.sol.retcode) retcode at profile point x = $(profiler_cache.x_cur). Profiling is interrupted."
+    
+  end
+  solver_cache.retcode_original = integrator.sol.retcode
+  return nothing
+end
+
+################################################## OPTIMIZATION PROFILER STEPPERS ##################################################
 
 const DEFAULT_INIT_STEP = 1e-2
 
@@ -105,90 +173,82 @@ prepare_initial_step(step::Function) = step
 get_step(s::AbstractProfilerStep{S}, pars, i) where S <: Function = s.initial_step(pars, i)
 
 
-function propose_next_pars!(profiler_state::ProfilerState, s::FixedStep)
[…]
