### 2025-09-29 ee427cc
func and pred tested on Boehm
