### 2025-09-30 723d433
func tests added
