+++ test/test_taxol_model.jl
@@ -40,13 +40,13 @@ profile_range = [
 plprob = ProfileLikelihoodProblem(optprob, p0; profile_lower = first.(profile_range), profile_upper = last.(profile_range), threshold = sigmasq*chi2_quantile(0.95, 5))
 
 
-@testset "Taxol model. Fixed-step OptimizationProfiler with derivative-free optimizer" begin
+@testset "Taxol model. Fixed-step OptimizationProfiler with gradient-basedoptimizer" begin
   
   idxs = 1:5
-  profile_step(p0, i) = p0[i] * 0.1
+  profile_step(p0, i) = p0[i] * 0.05
   atol = [profile_step(p0, i)/2 for i in idxs]
   atol[3] = 0.041 # tmp fix as r0 upper bound fails to be within step/2 tolerance
-  method = OptimizationProfiler(optimizer = NLopt.LN_NELDERMEAD(), stepper = FixedStep(; initial_step=profile_step))
+  method = OptimizationProfiler(optimizer = Optimization.LBFGS(), stepper = FixedStep(; initial_step=profile_step))
   
   #=
   method = OptimizationProfiler(optimizer = NLopt.LN_NELDERMEAD(), 
