### 2025-10-01 f485aab
gamma check fixed to allow gamma=0
