+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"profile_methods/#Profile-likelihood-methods","page":"Profile likelihood methods","title":"Profile likelihood methods","text":"","category":"section"},{"location":"profile_methods/","page":"Profile likelihood methods","title":"Profile likelihood methods","text":"LikelihoodProfiler provides a range of methods to profile likelihood functions and explore practical identifiability. The method should be provided as the second argument to the solve function.","category":"page"},{"location":"profile_methods/#Optimization-based-profiles","page":"Profile likelihood methods","title":"Optimization-based profiles","text":"","category":"section"},{"location":"profile_methods/","page":"Profile likelihood methods","title":"Profile likelihood methods","text":"The method computes profiles for each parameter by iteratively changing the value of the parameter and re-optimizing the likelihood function with respect to all other parameters. ","category":"page"},{"location":"profile_methods/#LikelihoodProfiler.OptimizationProfiler-profile_methods","page":"Profile likelihood methods","title":"LikelihoodProfiler.OptimizationProfiler","text":"OptimizationProfiler{S, opType, optsType}\n\nA profiler method that uses stepwise re-optimization to profile the likelihood function.\n\nFields\n\nstepper::S: The algorithm used to compute the next profile point. Supported steppers include:\nFixedStep: Proposes a constant step size in the profiling direction (Default). \nLineSearchStep: Uses a line search to adaptively determine the step size in the direction which is chosen by the direction keyword argument.\noptimizer::opType: The optimizer used for the optimization process.\noptimizer_opts::optsType: Options for the optimizer. Defaults to NamedTuple().\n\nStepping Options\n\nThe stepper argument controls how the next profile point is chosen. For example:\n\nstepper = FixedStep(initial_step=0.1): Use a constant step size of 0.1.\nstepper = LineSearchStep(direction=:Secant, linesearch=InterpolationLineSearch()): Use a line search with secant direction.\n\nSee the documentation for each stepper type (e.g., ?FixedStep, ?LineSearchStep) for more details and customization options.\n\nExample\n\nusing Optimization\nprofiler = OptimizationProfiler(; optimizer = Optimization.LBFGS(), optimizer_opts = (reltol=1e-4,))\n\n\n\n\n\n","category":"type"},{"location":"profile_methods/#Integration-based-profiles","page":"Profile likelihood methods","title":"Integration-based profiles","text":"","category":"section"},{"location":"profile_methods/","page":"Profile likelihood methods","title":"Profile likelihood methods","text":"The method computes profiles for each parameter (or function of parameters) by integrating the differential equations system. ","category":"page"},{"location":"profile_methods/#LikelihoodProfiler.IntegrationProfiler-profile_methods","page":"Profile likelihood methods","title":"LikelihoodProfiler.IntegrationProfiler","text":"IntegrationProfiler{opType, optsType, DEAlg, DEOpts}\n\nA profiler method that uses integration of differential equations system to profile the likelihood function.\n\nFields\n\nreoptimize::Bool: Indicates whether to re-optimization after each step of the integrator. Defaults to false.\noptimizer::opType: The optimizer used for the optimization process. Defaults to nothing.\noptimizer_opts::optsType: Options for the optimizer. Defaults to NamedTuple().\nintegrator::DEAlg: The differential equation algorithm used for integration.\nintegrator_opts:
[…]
