### 2025-10-02 c070a76
fix bounds idx
