### 2025-10-02 f479c28
add gradient based solver to taxol test
