### 2025-10-02 6fe6c35
fix log message in CICO
