+++ test/test_taxol_model.jl
@@ -28,21 +28,20 @@ function test_taxol(sol, i; kwargs...)
   taxol_retcodes[i][2] == :Identifiable && (@test isapprox(ci[2], taxol_ci[i][2]; kwargs...))
 end
 
+lb = [2.0, 2.0, 0.01, 0.05, 30.]
+ub = [30.0, 30.0, 0.6, 5.0, 250.0]
+
 optf = OptimizationFunction(taxol_obj, Optimization.AutoForwardDiff())
-optprob = OptimizationProblem(optf, p0)
+optprob = OptimizationProblem(optf, p0; lb=lb, ub=ub)
 
-profile_lower = [2.0, 2.0, 0.01, 0.05, 30.]
-profile_upper = [30.0, 30.0, 0.6, 5.0, 250.0]
-
-plprob = ProfileLikelihoodProblem(optprob, p0; profile_lower, profile_upper, threshold = sigmasq*chi2_quantile(0.95, 5))
+plprob = ProfileLikelihoodProblem(optprob, p0; threshold = sigmasq*chi2_quantile(0.95, 5))
 
 
 @testset "Taxol model. Fixed-step OptimizationProfiler with gradient-based optimizer" begin
 
   idxs = 1:5
-  profile_step(p0, i) = p0[i] * 0.01
-  atol = [0.2 for i in idxs]
-  #atol[3] = 0.041 # tmp fix as r0 upper bound fails to be within step/2 tolerance
+  profile_step(p0, i) = p0[i] * 0.1
+  atol = [profile_step(p0, i)/2 for i in idxs]
   method = OptimizationProfiler(; optimizer = Optimization.LBFGS(), stepper = FixedStep(; initial_step=profile_step))
   
   #=
