+++ docs/src/case_studies/taxol.md
@@ -136,15 +136,15 @@ sigmasq = (LikelihoodProfiler.mean([(Cerr005/C005); (Cerr010/C010); (Cerr040/C04
 Next, we construct the profile likelihood problem `ProfileLikelihoodProblem` and run the profiler for the five parameters:
 
 ```julia
+lb = [2.0, 2.0, 0.01, 0.05, 30.]
+ub = [30.0, 30.0, 0.6, 5.0, 250.0]
+
 optf = OptimizationFunction(taxol_obj, Optimization.AutoForwardDiff())
-optprob = OptimizationProblem(optf, p0)
+optprob = OptimizationProblem(optf, p0; lb=lb, ub=ub)
 
-profile_lower = [2.0, 2.0, 0.01, 0.05, 30.]
-profile_upper = [30.0, 30.0, 0.6, 5.0, 250.0]
+plprob = ProfileLikelihoodProblem(optprob, p0; threshold = sigmasq*chi2_quantile(0.95, 5))
 
-plprob = ProfileLikelihoodProblem(optprob, p0; profile_lower, profile_upper, threshold = sigmasq*chi2_quantile(0.95, 5))
-
-profile_step(p0, i) = p0[i] * 0.01
+profile_step(p0, i) = p0[i] * 0.1
 method = OptimizationProfiler(optimizer = Optimization.LBFGS(), stepper = FixedStep(; initial_step=profile_step))
 sol = solve(plprob, method)
 plot(sol, size=(800,300), margins=5Plots.mm)
