+++ Project.toml
@@ -5,7 +5,7 @@ license = "MIT"
 desc = "Practical identifiability analysis and confidence intervals estimation"
 authors = ["ivborissov <iv.borissov@gmail.com>", "metelkin <evgeny.metelkin@gmail.com>"]
 repo-url = "https://github.com/insysbio/LikelihoodProfiler.jl"
-version = "1.2.0"
+version = "1.2.1"
 
 [deps]
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
