### 2025-10-24 c9385f2
CompatHelper: bump compat for OptimizationBase to 4, (keep existing compat)
