+++ Project.toml
@@ -34,7 +34,7 @@ Distributed = "1.10"
 Distributions = "0.25"
 ForwardDiff = "0.10.36, 1"
 Optimization = "4.1.1"
-OptimizationBase = "2.4.0"
+OptimizationBase = "2.4.0, 4"
 OptimizationNLopt = "0.3.1"
 PreallocationTools = "0.4"
 RecipesBase = "1.1"
