### 2025-11-06 #8929 @lockwo
The description of the software is there, it's within scope, and it has a OSI approved license, so we can assemble the reviewers now.

@ivborissov Please suggest ~5 potential reviewers. You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission. Or people who aren't in the JOSS system would also be ok. If you know their GitHub usernames, then please say them here in this thread (without tagging them with an @). You can also investigate the "AI"-suggested papers and their authors located in the previous posts from our editorial bot.
