### 2025-11-24 #9501 @editorialbot
[REVIEW]: LikelihoodProfiler.jl: Unified profile-likelihood workflows for identifiability and confidence intervals
**Submitting author:** @ivborissov (<a href="http://orcid.org/0000-0002-2887-1903">Ivan Borisov</a>)
**Repository:** https://github.com/insysbio/LikelihoodProfiler.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.3
