### 2025-11-24 ad40b3c
upd paper with bands
