+++ joss/paper.md
@@ -27,84 +27,82 @@ bibliography: paper.bib
 
 ## Summary
 
-Practical identifiability addresses how well a mechanistic model is determined by available experimental data. Profile likelihood-based methods are widely used to determine practical identifiability and serve as a proxy for structural identifiability analysis, particularly when the complexity of a model makes structural methods impractical [@Heinrich2025]. Profile likelihood techniques can be extended beyond parameter analysis to assess the identifiability of model states and predictions. This versatility makes profile likelihood analysis an essential component in the development and validation of models.
+Practical identifiability addresses how well a mechanistic model is determined by available experimental data. Profile likelihood-based methods are widely used to explore practical identifiability and serve as a proxy for structural identifiability analysis [@Heinrich2025]. Moreover, these methods can be extended beyond parameter analysis to assess the identifiability of model states and predictions. This versatility makes profile likelihood an essential component in the development and validation of models.
 
-`LikelihoodProfiler.jl` is an open-source Julia package designed to perform profile likelihood-based identifiability analyses by offering a unified and extensible interface.
+The paper presents `LikelihoodProfiler.jl`, an open-source Julia package for conducting profile likelihood-based identifiability and predictability analysis.
 
 ## Statement of Need
 
-Despite the widespread use of profile likelihood methods in practical identifiability analysis, existing tools often lack a common interface and extensive ecosystem integration, limiting accessibility and reproducibility. Also different profile likelihood based methods are implemented in different software tools, which requires the user to switch between languages and software interfaces. `LikelihoodProfiler.jl` addresses these limitations by providing:
-- Unified interface to access multiple profiling methods.
-- Compatibility with common modeling standards (Heta [@Metelkin2021], PEtab [@Persson2025], SBML).
+Despite the widespread use of profile likelihood methods for practical identifiability analysis, existing tools are typically bound to specific modeling frameworks or software environments. Moreover, different profile likelihood algorithms are implemented in separate tools, which requires the user to switch between languages and software interfaces. `LikelihoodProfiler.jl` addresses these limitations by providing:
+- Unified interface to access multiple profile likelihood algorithms.
+- Compatibility with common modeling standards (SBML, PEtab [@Persson2025], Heta [@Metelkin2021]).
 - Integration with Julia’s SciML [@Rackauckas2017] for efficient computation and extensibility.
 
 ## Features and Methodologies
 
-`LikelihoodProfiler.jl` provides a unified interface for various profile likelihood methods, including optimization-based (**OptimizationProfiler**) and integration-based profiles (**IntegrationProfiler**), CI endpoints search (**CICOProfiler**), and more.
+`LikelihoodProfiler.jl` provides a unified interface for various profile likelihood methods, including optimization-based **OptimizationProfiler** and integration-based profiles **IntegrationProfiler**, CI endpoints search **CICOProfiler** [@Borisov2020], and more.
 
 All methods leverage a CommonSolve interface [@Rackauckas2017] (`CommonSolve.solve()`), supporting global settings for parallelization and verbosity. Profiling results are directly visualizable using the `Plots.jl` package and exportable as DataFrames for further analysis.
 
 ## Demonstrative Example: JAK/STAT Signaling Pathway Model
 
[…]
