+++ joss/ci_table.md
@@ -1,11 +1,0 @@
-| **parameter**<br>`Symbol`     | **OptimizationProfiler**<br>`Tuple{Any, Any}` | **IntegrationProfiler**<br>`Tuple{Any, Any}` | **CICOProfiler**<br>`Tuple{Any, Any}` |
-|------------------------------:|----------------------------------------------:|---------------------------------------------:|--------------------------------------:|
-| log10\_Epo\_degradation\_BaF3 | (-1.79612, -1.33173)                          | (-1.7975, -1.33177)                          | (-1.79661, -1.3302)                   |
-| log10\_k\_exp\_hetero         | (nothing, -2.65263)                           | (nothing, -2.65244)                          | (nothing, -2.65213)                   |
-| log10\_k\_exp\_homo           | (-2.68323, -1.73872)                          | (-2.68658, -1.73878)                         | (-2.68367, -1.73888)                  |
-| log10\_k\_imp\_hetero         | (-1.94661, -1.62555)                          | (-1.94726, -1.62151)                         | (-1.94888, -1.62229)                  |
-| log10\_k\_imp\_homo           | (-0.109411, nothing)                          | (-0.109301, nothing)                         | (-0.109678, nothing)                  |
-| log10\_k\_phos                | (4.08355, 4.33082)                            | (4.07961, 4.3315)                            | (4.07836, 4.33194)                    |
-| log10\_sd\_pSTAT5A\_rel       | (0.364585, 0.870501)                          | (0.363312, 0.869364)                         | (0.362142, 0.870803)                  |
-| log10\_sd\_pSTAT5B\_rel       | (0.625924, 1.08439)                           | (0.627294, 1.08405)                          | (0.62272, 1.0853)                     |
-| log10\_sd\_rSTAT5A\_rel       | (0.312047, 0.762828)                          | (0.311183, 0.762745)                         | (0.307908, 0.763785)                  |
