+++ joss/paper.md
@@ -27,42 +27,43 @@ bibliography: paper.bib
 
 ## Summary
 
-Practical identifiability addresses how well a mechanistic model is determined by available experimental data. Profile likelihood-based methods are widely used to explore practical identifiability and serve as a proxy for structural identifiability analysis [@Heinrich2025]. Moreover, these methods can be extended beyond parameter analysis to assess the identifiability of model states and predictions. This versatility makes profile likelihood an essential component in the development and validation of models.
+Practical identifiability addresses how well a mechanistic model is determined by available experimental data. Profile likelihood-based methods are widely used to explore practical identifiability and often serve as a proxy for structural identifiability analysis [@Heinrich2025]. Moreover, these methods can also be generalized to states and predictions analysis. This versatility makes profile likelihood an essential component in model development.
 
 The paper presents `LikelihoodProfiler.jl`, an open-source Julia package for conducting profile likelihood-based identifiability and predictability analysis.
 
 ## Statement of Need
 
-Despite the widespread use of profile likelihood methods for practical identifiability analysis, existing tools are typically bound to specific modeling frameworks or software environments. Moreover, different profile likelihood algorithms are implemented in separate tools, which requires the user to switch between languages and software interfaces. `LikelihoodProfiler.jl` addresses these limitations by providing:
-- Unified interface to access multiple profile likelihood algorithms.
-- Compatibility with common modeling standards (SBML, PEtab [@Persson2025], Heta [@Metelkin2021]).
+Current profile likelihood implementations are often tied to specific modeling frameworks or software environments. In addition, different profile likelihood algorithms are spread across separate tools, forcing users to switch between languages and interfaces.
+`LikelihoodProfiler.jl` addresses these limitations by providing:
+- Unified interface to multiple profile likelihood algorithms.
+- Compatibility with common modeling standards (Heta [@Metelkin2021], SBML, PEtab [@Persson2025]).
 - Integration with Julia’s SciML [@Rackauckas2017] for efficient computation and extensibility.
 
 ## Features and Methodologies
 
-`LikelihoodProfiler.jl` provides a unified interface for various profile likelihood methods, including optimization-based **OptimizationProfiler** and integration-based profiles **IntegrationProfiler**, CI endpoints search **CICOProfiler** [@Borisov2020], and more.
+`LikelihoodProfiler.jl` provides a unified interface for various profile likelihood methods, including optimization-based **OptimizationProfiler** and integration-based profiles **IntegrationProfiler**, CI endpoints search **CICOProfiler**, and more.
 
-All methods leverage a CommonSolve interface [@Rackauckas2017] (`CommonSolve.solve()`), supporting global settings for parallelization and verbosity. Profiling results are directly visualizable using the `Plots.jl` package and exportable as DataFrames for further analysis.
+All methods leverage a `CommonSolve` interface [@Rackauckas2017] (`CommonSolve.solve()`), supporting global settings for parallelization and verbosity. Profiling results are directly visualizable using the `Plots.jl` package and exportable as DataFrames for further analysis.
 
 ## Demonstrative Example: JAK/STAT Signaling Pathway Model
 
[…]
