+++ joss/joss.jl
@@ -1,4 +1,4 @@
-using Optimization, LikelihoodProfiler, OrdinaryDiffEq, Plots, PEtab
+using Optimization, LikelihoodProfiler, OrdinaryDiffEq, Plots, PEtab, CICOBase
 
 model_name = "Boehm_JProteomeRes2014"
 path_yaml = joinpath(@__DIR__, "../models/", "$model_name/$model_name.yaml")
@@ -15,38 +15,37 @@ alg = IntegrationProfiler(integrator = Tsit5(), integrator_opts = (dtmax=0.5, re
 
 sol_param = solve(param_profile_prob, alg, verbose=true)
 
+alg_cico = CICOProfiler(optimizer = :LN_NELDERMEAD, scan_tol = 1e-10)
+sol_cico = solve(param_profile_prob, alg_cico, verbose=true)
+
 w, h = 1000, 300
 
-p11 = plot(sol1[1] , dpi=400, xguide="log10_Epo_degradation_BaF3", yguide="likelihood", legend=false)
-p12 = plot(sol1[2] , dpi=400, xguide="log10_k_exp_hetero", yguide="likelihood", legend=false, title="Parameter Profiles")
-p13 = plot(sol1[3] , dpi=400, xguide="log10_k_exp_homo", yguide="likelihood", legend=:outerright)
+p11 = plot(sol_param[1] , dpi=400, xguide="log10_Epo_degradation_BaF3", yguide="likelihood", legend=false)
+p12 = plot(sol_param[2] , dpi=400, xguide="log10_k_exp_hetero", yguide="likelihood", legend=false, title="Parameter Profiles")
+p13 = plot(sol_param[3] , dpi=400, xguide="log10_k_exp_homo", yguide="likelihood", legend=:outerright)
 
 p1 = plot(p11, p12, p13, layout=(1,3), size=(w,h), dpi=400)
 
-p21 = plot(sol2[1] , dpi=400, xguide="log10_Epo_degradation_BaF3", yguide="likelihood", legend=false)
-p22 = plot(sol2[2] , dpi=400, xguide="log10_k_exp_hetero", yguide="likelihood", legend=false, title="IntegrationProfiler")
-p23 = plot(sol2[3] , dpi=400, xguide="log10_k_exp_homo", yguide="likelihood", legend=:outerright)
+savefig(p1, joinpath(@__DIR__, "param_profiles.png"))
+
+p21 = plot(sol_cico[1] , dpi=400, xguide="log10_Epo_degradation_BaF3", yguide="likelihood", legend=false)
+p22 = plot(sol_cico[2] , dpi=400, xguide="log10_k_exp_hetero", yguide="likelihood", legend=false, title="CICOProfiler (Confidence Intervals endpoints)")
+p23 = plot(sol_cico[3] , dpi=400, xguide="log10_k_exp_homo", yguide="likelihood", legend=:outerright)
 
 p2 = plot(p21, p22, p23, layout=(1,3), size=(w,h), dpi=400)
 
-p31 = plot(sol3[1] , dpi=400, xguide="log10_Epo_degradation_BaF3", yguide="likelihood", legend=false)
-p32 = plot(sol3[2] , dpi=400, xguide="log10_k_exp_hetero", yguide="likelihood", legend=false, title="CICOProfiler")
-p33 = plot(sol3[3] , dpi=400, xguide="log10_k_exp_homo", yguide="likelihood", legend=:outerright)
+savefig(p2, joinpath(@__DIR__, "cico_profiles.png"))
 
-p3 = plot(p31, p32, p33, layout=(1,3), size=(w,h), dpi=400)
-
[…]
