### 2025-11-24 #9501 @lockwo
> [@lockwo](https://github.com/lockwo) Thank you for starting the review process. Before the reviewers begin, I would like to ask whether updates to the manuscript are still allowed at this stage. I am ready to commit the revised text and create a new tagged release if this is acceptable.

Yes, you can update the text (and the code and paper will probably be iterated on through this review), just run `@editorialbot generate pdf` in this thread after doing so, that way we know there's a new copy.
