### 2025-11-24 0b3471f
tests and readme updated
