### 2025-11-28 insysbio/LikelihoodProfiler.jl#70 @salbalkus
JOSS Review: Reproducibility
Raising as part of JOSS Review openjournals/joss-reviews#9501

Great job on providing relevant examples! Most of the code I was able to run effectively. That said, several minor reproducibility issues arose:

- In both the paper and the documentation, when I ran the code directly in a fresh project, any lines that called `Optimization.LBFGS()` threw the error `UndefVarError: LBFGS not defined in Optimization`. I speculate this error arises because the environments used to generate the paper use an old version of the Optimization package and call functionality that has been removed or renamed. Given that most users will likely be trying to use the most recent versions of Optimization, I think the code examples should be updated to ensure compatibility.
- When I ran `joss.jl` in the `joss` environment, the last line threw `PreparationMismatchError (inconsistent types between preparation and execution)`
- When I ran the Taxol example in the documentation, the second code block threw  `UndefVarError: mean not defined in LikelihoodProfiler`
- The third code box in the JOSS paper has a line of code that runs off the page, so readers cannot fully see the parameters being specified.

If these issues can be addressed, I should be able to check off the Reproducibility blocker on my review. Thanks!
