### 2025-11-28 insysbio/LikelihoodProfiler.jl#71 @salbalkus
JOSS Review: Documentation and Units Tests
Raising as part of JOSS Review openjournals/joss-reviews#9501\

Nice work setting up a tutorial in Documenter and a unit testing suite. Just a few small issues to be raised with the documentation before I can check off this section:

- As a statistician, I'm familiar with the notion of profile likelihood in the context of statistical modeling. However, this package seems mostly targeted at ODE solvers and the field of computational or systems biology, which is a quite different audience. As a "Statement of Need" in the landing page of the documentation, I would like to see more of a clarification of who the *target audience* is and what kinds of applied problems they might expect to solve using this package.
- Some of the plot labels are a little unclear to me. In the tutorial, it would be great to see a more in-depth description of the plot's output (for example, explaining the profile steps, threshold, how the likelihood curve is constructed, etc.)
- In the unit tests,  `test_jak-stat_model.jl` is commented out and test_profile.jl and `test_adaptive`.jl seems to not be called. Can these be safely removed if they are not longer being used?

Thank you for the thorough examples and plots!
