### 2025-11-28 insysbio/LikelihoodProfiler.jl#72 @salbalkus
JOSS Review: Paper
Raising as part of JOSS Review openjournals/joss-reviews#9501

Overall, I thought the paper was nicely written. Before I can check it off on my review, a few issues surrounding clarity ought to be addressed:
- The description of the high-level functionality and purpose of the software needs to be targeted at a non-specialist audience. This will require a more basic background of the methods: for example, defining what a profile likelihood is, what a mechanistic model is,  what is "practical identifiability" etc. so that a person outside systems biology who does not use these methods normally can understand what the package does.
- Regarding the statement of need, I found it somewhat vague. It should be providing a high-level description of what the package's functions do, and what kinds of applied problems it can solve, so a user will know whether to use it or not. I think the current content, which focuses more on issues with existing packages, would fit better in the "state of the field" part of the paper as a contrast of how LikelihoodProfiler.jl differs from what's out there now.
- Minor formatting issues:
	- In a few sections of the paper, it looks like Markdown bullets points have broken and turned into dashes -- be sure to put a line break between the main text and the bulleted list in the Markdown.
	- PEtab.jl citation is missing the journal / source.
	- PEtab Benchmark citation: I think "contributors" should be capitalized.
