### 2025-12-04 26a9218
fix mean function use in taxol
