+++ docs/src/case_studies/taxol.md
@@ -4,7 +4,7 @@ As an example of practical identifiability analysis, we use the **Cancer Taxol T
 
 ```julia
 using LikelihoodProfiler, Test
-using OptimizationLBFGSB, ForwardDiff, OrdinaryDiffEq
+using OptimizationLBFGSB, ForwardDiff, OrdinaryDiffEq, Distributions
 
 # https://github.com/marisae/cancer-chemo-identifiability/blob/master/Profile%20Likelihood/testa0_de.m
 function ode_func(du, u, p, t, drug)
@@ -70,10 +70,10 @@ Cerr040 = [0.005, 0.010, 0.010, 0.011, 0.010, 0.010]*1091.0   # thousands of cel
 Cell100 = [0.009, 0.025, 0.026, 0.028, 0.029, 0.031]*1091.0   # thousands of cells
 Cerr100 = [0.006, 0.010, 0.009, 0.008, 0.011, 0.011]*1091.0   # thousands of cells
 
-C005 = LikelihoodProfiler.mean(Cell005)
-C010 = LikelihoodProfiler.mean(Cell010)
-C040 = LikelihoodProfiler.mean(Cell040)
-C100 = LikelihoodProfiler.mean(Cell100)
+C005 = mean(Cell005)
+C010 = mean(Cell010)
+C040 = mean(Cell040)
+C100 = mean(Cell100)
 
 data = [Cell005/C005, Cell010/C010, Cell040/C040, Cell100/C100]
 datamean = [C005, C010, C040, C100]
@@ -130,7 +130,7 @@ end
 
 # threshold is chosen according to
 # https://github.com/marisae/cancer-chemo-identifiability/blob/master/Profile%20Likelihood/testa0_fit.m#L40-L41
-sigmasq = (LikelihoodProfiler.mean([(Cerr005/C005); (Cerr010/C010); (Cerr040/C040); (Cerr100/C100)]))^2
+sigmasq = (mean([(Cerr005/C005); (Cerr010/C010); (Cerr040/C040); (Cerr100/C100)]))^2
 ```
 
 Next, we construct the profile likelihood problem `ProfileLikelihoodProblem` and run the profiler for the five parameters:
