### 2025-12-04 3255e42
add ADTypes to deps
