+++ src/LikelihoodProfiler.jl
@@ -4,6 +4,7 @@ using SciMLBase, PreallocationTools
 using SimpleUnPack: @unpack
 using Reexport
 @reexport import SciMLBase: OptimizationFunction, OptimizationProblem, remake, solve, solve!, init
+@reexport using ADTypes
 @reexport using DataFrames
 using LinearAlgebra, DataInterpolations
 using Distributions: quantile, Chisq
