+++ joss/paper.bib
@@ -77,7 +77,7 @@
   doi = {10.1021/pr5006923}
 }
 
-@software{vaibhav_kumar_dixit_2023_7738525,
+@software{Dixit2023,
 	author = {Vaibhav Kumar Dixit and Christopher Rackauckas},
 	month = mar,
 	publisher = {Zenodo},
