### 2025-12-04 9723233
OptimizationBase 2.4 added to work with PEtab
