+++ Project.toml
@@ -35,7 +35,7 @@ DataInterpolations = "6.5.2, 7"
 Distributed = "1.10"
 Distributions = "0.25"
 ForwardDiff = "0.10.36, 1"
-OptimizationBase = "3, 4"
+OptimizationBase = "2.4, 3, 4"
 OptimizationLBFGSB = "1.2.1"
 OptimizationNLopt = "0.3.1"
 PreallocationTools = "0.4"
