### 2025-12-04 f5c907c
Pkg overview updated in the docs
