### 2025-12-05 6aa3ecf
objective function label
