+++ docs/src/case_studies/taxol.md
@@ -2,9 +2,8 @@
 
 As an example of practical identifiability analysis, we use the **Cancer Taxol Treatment Model**. It is an ODE model with 3 state variables and 5 parameters. The identifiability of this model was studied in [Marisa C.Eisenberg, Harsh V.Jain. A confidence building exercise in data and identifiability](https://www.sciencedirect.com/science/article/pii/S0022519317303454). We have translated [author's MATLAB code](https://github.com/marisae/cancer-chemo-identifiability) into Julia. The model is defined by the following system of differential equations:
 
-```julia
-using LikelihoodProfiler, Test
-using OptimizationLBFGSB, ForwardDiff, OrdinaryDiffEq, Distributions
+```@example taxol-1
+using LikelihoodProfiler, OptimizationLBFGSB, OrdinaryDiffEq, Distributions
 
 # https://github.com/marisae/cancer-chemo-identifiability/blob/master/Profile%20Likelihood/testa0_de.m
 function ode_func(du, u, p, t, drug)
@@ -39,7 +38,7 @@ end
 
 Experimental datasets are also provided in the cancer-chemo-identifiability repo for four drug doses (5, 10, 40, 100)
 
-```julia
+```@example taxol-1
 # https://github.com/marisae/cancer-chemo-identifiability/blob/master/Profile%20Likelihood/testa0_fit.m
 
 # Data from Terzis et al. Brit J Cancer 1997;75:1744.
@@ -81,7 +80,7 @@ datamean = [C005, C010, C040, C100]
 
 Next we define solver options, initial values, optimal parameter values, and tspan
 
-```julia
+```@example taxol-1
 # solver algorithm and tolerances
 solver_opts = Dict(
     :alg => AutoTsit5(Rosenbrock23()),
@@ -99,20 +98,12 @@ p0 = [8.3170, 8.0959, 0.0582, 1.3307, 119.1363]
 tspan = (0.,15.)
 ```
  
- We use OLS for the objective function as proposed in the original code.
+ We define an objective function (OLS) as in the original publication. The profile likelihood evaluates how this objective changes when one parameter is varied and all other parameters are re-optimized.
 
- ```julia
+ ```@example taxol-1
 # https://github.com/marisae/cancer-chemo-identifiability/blob/master/Profile%20Likelihood/testa0_fit.m#L92
 # https://www.mathworks.com/help/optim/ug/lsqcurvefit.html
-function taxol_obj(
-  x, _p;
-  ode_func=ode_func,
-  dose=dose,
-  data=data,
-  datamean=datamean,
-  times=times,
-  solver_opts=solver_opts
-)
+function taxol_obj(x, _p)
   loss = 0.
   for (i,d) in enumerate(dose)
      prob = ODEProblem((du,u,p,t)->ode_func(du,u,p,t,d), u0, tspan, x)
@@ -127,29 +118,43 @@ function taxol_obj(
   end
   return loss
 end
+```
 
+We define the threshold as in the original publication. The threshold line represents the value of the objective function (or negative log-likelihood) that corresponds to the chosen confidence level. Where the profile curve intersects this threshold gives the lower and upper CI bounds.
+
+```@example taxol-1
 # threshold is chosen according to
 # https://github.com/marisae/cancer-chemo-identifiability/blob/master/Profile%20Likelihood/testa0_fit.m#L40-L41
 sigmasq = (mean([(Cerr005/C005); (Cerr010/C010); (Cerr040/C040); (Cerr100/C100)]))^2
+threshold = sigmasq*chi2_quantile(0.95, 5)
 ```
 
-Next, we construct the profile likelihood problem `ProfileLikelihoodProblem` and run the profiler for the five parameters:
[…]
