+++ docs/Project.toml
@@ -2,16 +2,11 @@
 CICOBase = "33444335-3134-4342-4659-594331344435"
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-OptimizationNLopt = "4e6fcdb7-1186-4e1f-a706-475e75c168bb"
+OptimizationLBFGSB = "22f7324a-a79d-40f2-bebe-3af60c77bd15"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 
 [compat]
 CICOBase = "0.6.0"
-Distributed = "1.10"
 Documenter = "1"
-ForwardDiff = "0.10.36, 1"
-OptimizationNLopt = "0.3.1"
 OrdinaryDiffEq = "6"
-Plots = "1"
