### 2025-12-08 c841269
Add Plots to Taxol tutorial
