### 2025-12-08 7adac77
Fix @ref profile_likelihood_methods
