### 2025-12-08 581f231
Fix rosenbrock ref
