### 2025-12-08 c6906c5
change ODE IntegrationProfiler solver for SIR model
