+++ test/test_sir_model.jl
@@ -56,7 +56,7 @@ end
   
   idxs = 1:3
   rtol = 3e-3 # how to set it?
-  method = IntegrationProfiler(integrator = Rodas5P(autodiff = AutoFiniteDiff()), matrix_type = :hessian)
+  method = IntegrationProfiler(integrator = FBDF(autodiff = AutoFiniteDiff()), matrix_type = :hessian)
   sol = solve(plprob, method)
   for i in idxs
     test_sir(sol, i; rtol = rtol)
