### 2025-12-08 fa9e042
Fix hasoptimizer
