### 2025-12-08 c96967e
FIx get_optimizer and get_optimizer_opts
