### 2025-12-08 c011466
SIR model updated
