### 2025-12-08 27389d5
gaussian ex added to Tutorials
