### 2025-12-09 0602864
addind details to Getting started docs
