+++ joss/paper.md
@@ -27,46 +27,41 @@ bibliography: paper.bib
 
 ## Summary
 
-Practical identifiability addresses how well a mechanistic model is determined by available experimental data. Profile likelihood-based methods are widely used to explore practical identifiability and often serve as a proxy for structural identifiability analysis [@Heinrich2025]. Moreover, these methods can also be generalized to states and predictions analysis. This versatility makes profile likelihood an essential component in model development.
+Mathematical models — mechanistic, statistical, or otherwise — contain unknown parameters that must be estimated from data to enable reliable predictions. Profile likelihood is a widely used approach for studying the *practical identifiability* of such models, that is, assessing how well parameters are determined by available observations [@Heinrich2025]. The method recovers the local shape of the objective function around the optimum and provides confidence intervals that quantify uncertainty in parameter estimates. Moreover, profile likelihood techniques can be generalized to analyze derived quantities such as model states and predictions. This versatility makes profile likelihood an essential component in model development and uncertainty assessment.
 
-The paper presents `LikelihoodProfiler.jl`, an open-source Julia package for conducting profile likelihood-based identifiability and predictability analysis.
+`LikelihoodProfiler.jl` is an open-source Julia package that implements several profile likelihood methods through a unified interface, enabling parameters and prediction uncertainty analysis across general maximum-likelihood estimation problems.
 
 ## Statement of Need
 
-Current profile likelihood implementations are often tied to specific modeling frameworks or software environments. In addition, different profile likelihood algorithms are spread across separate tools, forcing users to switch between languages and interfaces.
-`LikelihoodProfiler.jl` addresses these limitations by providing:
-- Unified interface to multiple profile likelihood algorithms.
-- Compatibility with common modeling standards (Heta [@Metelkin2021], SBML, PEtab [@Persson2025]).
-- Integration with Julia’s SciML [@Rackauckas2017] for efficient computation and extensibility.
+Researchers across many scientific domains need tools to quantify how well model parameters and predictions are constrained by data. Profile likelihood methods provide an intuitive and interpretable approach to parameter and prediction uncertainty, but existing software implementations - such as `sbioparameterci (MATLAB)`, `dMod (R)`, and `pyPESTO (Python)` - are often tied to specific modeling ecosystems or limited to a single profiling method. Users working with statistical models, mechanistic models, or simulation-based models may therefore need to combine multiple tools or reimplement profiling routines for each workflow.
 
-## Features and Methodologies
-
-`LikelihoodProfiler.jl` provides a unified interface for various profile likelihood methods, including optimization-based **OptimizationProfiler** and integration-based profiles **IntegrationProfiler**, CI endpoints search **CICOProfiler**, and more.
-
-All methods leverage a `CommonSolve` interface [@Rackauckas2017] (`CommonSolve.solve()`), supporting global settings for parallelization and verbosity. Profiling results are directly visualizable using the `Plots.jl` package and exportable as DataFrames for further analysis.
[…]
