+++ joss/paper.bib
@@ -11,13 +11,13 @@
 }
 
 @misc{petab_benchmark_collection,
-  author       = {The PEtab Benchmark Collection contributors},
-  title        = {{The PEtab Benchmark Collection of parameter estimation problems}},
-  year         = 2024,
-  publisher    = {Zenodo},
-  version      = {2024.10.15},
-  doi          = {10.5281/zenodo.8155057},
-  url          = {https://doi.org/10.5281/zenodo.8155057}
+  author = {The PEtab Benchmark Collection Contributors},
+  title = {{The PEtab Benchmark Collection of parameter estimation problems}},
+  year = 2024,
+  publisher = {Zenodo},
+  version = {2024.10.15},
+  doi = {10.5281/zenodo.8155057},
+  url = {https://doi.org/10.5281/zenodo.8155057}
 }
 
 @article{Metelkin2021,
@@ -84,18 +84,19 @@
 	title = {Optimization.jl: A Unified Optimization Package},
 	version = {v3.12.1},
 	doi = {10.5281/zenodo.7738525},
-  	url = {https://doi.org/10.5281/zenodo.7738525},
-	year = 2023}
+  url = {https://doi.org/10.5281/zenodo.7738525},
+	year = 2023
+}
 
-  @article{Kreutz2012,
-    author = {Kreutz, C. and Raue, A. and Timmer, J.},
-    title = {Likelihood Based Observability Analysis and Confidence Intervals for Predictions of Dynamic Models},
-    journal = {BMC Systems Biology},
-    year = {2012},
-    volume = {6},
-    number = {120},
-    pages = {1--12},
-    doi = {10.1186/1752-0509-6-120}
+@article{Kreutz2012,
+  author = {Kreutz, C. and Raue, A. and Timmer, J.},
+  title = {Likelihood Based Observability Analysis and Confidence Intervals for Predictions of Dynamic Models},
+  journal = {BMC Systems Biology},
+  year = {2012},
+  volume = {6},
+  number = {120},
+  pages = {1--12},
+  doi = {10.1186/1752-0509-6-120}
 }
 
 @article{Chen2002,
