### 2025-12-09 033f60c
Fix formulas in the tutorials
