+++ joss/Project.toml
@@ -5,3 +5,5 @@ OptimizationLBFGSB = "22f7324a-a79d-40f2-bebe-3af60c77bd15"
 OrdinaryDiffEqTsit5 = "b1df2697-797e-41e3-8120-5422d3b24e4a"
 PEtab = "48d54b35-e43e-4a66-a5a1-dde6b987cf69"
 
+[compat]
+LikelihoodProfiler = "1.3.0"
