+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v1.3",
   "v1.2",
   "v1.1",
   "v1.0",
@@ -8,5 +9,5 @@ var DOC_VERSIONS = [
   "v0.3",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.2.1";
+var DOCUMENTER_NEWEST = "v1.3.0";
 var DOCUMENTER_STABLE = "stable";
