### 2025-12-09 495ea7e
Summary updated
