### 2025-12-12 insysbio/LikelihoodProfiler.jl#70 @salbalkus
Thank for you for making these updates. Everything in the documentation and the paper now runs great!
