### 2025-12-16 bfffbe8
Paper edited
