+++ joss/paper.bib
@@ -33,12 +33,15 @@
 }
 
 @article{Persson2025,
-  author = {S. Persson and others},
-  title = {PEtab.jl: Advancing the Efficiency and Utility of Dynamic Modelling},
-  year = {2025},
-  month = {May},
-  note = {Preprint},
-  doi = {10.1101/2025.04.30.651378}
+  author  = {Sebastian Persson and Fabian Fr{\"o}hlich and Stephan Grein and Torkel Loman and Damiano Ognissanti and Viktor Hasselgren and Jan Hasenauer and Marija Cvijovic},
+  title   = {PEtab.jl: Advancing the efficiency and utility of dynamic modelling},
+  journal = {Bioinformatics},
+  volume  = {41},
+  number  = {9},
+  pages   = {btaf497},
+  year    = {2025},
+  doi     = {10.1093/bioinformatics/btaf497},
+  publisher = {Oxford University Press}
 }
 
 @article{Rackauckas2017,
