### 2025-12-18 543fb5c
pkg refs added
