+++ joss/paper.md
@@ -33,7 +33,7 @@ Mathematical models — mechanistic, statistical, or otherwise — contain unkno
 
 ## Statement of Need
 
-Researchers across many scientific domains need tools to quantify how well estimates of model parameters and predictions are constrained by data. Profile likelihood methods provide an intuitive and interpretable approach to parameter and prediction uncertainty, but existing software implementations - such as `sbioparameterci (MATLAB)`, `dMod (R)`, and `pyPESTO (Python)` - are often tied to specific modeling ecosystems or limited to a single profiling method.  Within the Julia ecosystem, packages such as `ProfileLikelihood.jl` and `InformationGeometry.jl` provide related functionality but focus on specific profile likelihood methods rather than offering multiple profiling strategies through a unified interface. As a result, users working with different model types may need to combine multiple tools or reimplement profiling routines.
+Researchers across many scientific domains need tools to quantify how well estimates of model parameters and predictions are constrained by data. Profile likelihood methods provide an intuitive and interpretable approach to parameter and prediction uncertainty, but existing software implementations - such as `sbioparameterci (MATLAB)` [@sbioparameterci], `dMod (R)` [@Kaschek2019], and `pyPESTO (Python)` [@Schalte2025] - are often tied to specific modeling ecosystems or limited to a single profiling method.  Within the Julia ecosystem, packages such as `ProfileLikelihood.jl` [@VandenHeuvel2021] and `InformationGeometry.jl` [@Arutjunjan2025] provide related functionality but focus on specific profile likelihood methods rather than offering multiple profiling strategies through a unified interface. As a result, users may need to combine multiple tools or reimplement profiling routines.
 
 `LikelihoodProfiler.jl` addresses this need by providing a unified, model-agnostic interface for computing profile likelihoods for parameters and arbitrary functions of parameters. The package offers several profiling methods - such as optimization-based `OptimizationProfiler`, integration-based `IntegrationProfiler`, confidence-interval endpoint search via `CICOProfiler` - and integrates with the Julia SciML [@Rackauckas2017], making profile likelihood analysis accessible and efficient for a wide variety of applied problems.
 
