+++ joss/paper.bib
@@ -112,3 +112,77 @@
   pages = {714--732},
   doi = {10.1198/106186002493}
 }
+
+@software{Arutjunjan2025,
+  author       = {Rafael Arutjunjan and
+                  Simon Beyer},
+  title        = {RafaelArutjunjan/InformationGeometry.jl: v1.22.3},
+  month        = dec,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {v1.22.3},
+  doi          = {10.5281/zenodo.17882740},
+  url          = {https://doi.org/10.5281/zenodo.17882740},
+}
+
+@software{Schalte2025,
+  author       = {Schälte, Yannik and
+                  Fröhlich, Fabian and
+                  Stapor, Paul and
+                  Vanhoefer, Jakob and
+                  Weindl, Daniel and
+                  Jost, Paul Jonas and
+                  Wang, Dantong and
+                  Lakrisenko, Polina and
+                  Raimúndez, Elba and
+                  Pathirana, Dilan and
+                  Schmiester, Leonard and
+                  Städter, Philipp and
+                  Contento, Lorenzo and
+                  Merkt, Simon and
+                  Dudkin, Erika and
+                  Grein, Stephan and
+                  Hasenauer, Jan},
+  title        = {pyPESTO - Parameter EStimation TOolbox for python},
+  month        = nov,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {v0.5.7},
+  doi          = {10.5281/zenodo.17603051},
+  url          = {https://doi.org/10.5281/zenodo.17603051},
+  swhid        = {swh:1:dir:0f23cef4612e59a104f16f2a98265e67cc303ae5
+                   ;origin=https://doi.org/10.5281/zenodo.2553546;vis
+                   it=swh:1:snp:e1b577f4a2ae214ed2593d2b59cdd8b99183f
+                   dce;anchor=swh:1:rel:349dcb5eabe351b037a53f8a110f8
+                   9015d2ca3f1;path=ICB-DCM-pyPESTO-8e03762
+                  },
+}
+
+@article{Kaschek2019,
+ title={Dynamic Modeling, Parameter Estimation, and Uncertainty Analysis in R},
+ volume={88},
+ url={https://www.jstatsoft.org/index.php/jss/article/view/v088i10},
+ doi={10.18637/jss.v088.i10},
[…]
