### 2025-12-18 insysbio/LikelihoodProfiler.jl#72 @ivborissov
@salbalkus Thanks for the suggestion — that makes sense. I’ve added explicit citations for the related packages mentioned in the Statement of Need so that they are included in the references.
