### 2025-12-18 insysbio/LikelihoodProfiler.jl#72 @salbalkus
Great! I've checked everything off on my end for checklist 👍
