### 2025-12-29 75d50ac
joss fig captions updated
