### 2025-12-29 insysbio/LikelihoodProfiler.jl#71 @ivborissov
@currocam Thank you for pointing this out. I’ve added a short Community Guidelines section to the README describing how to contribute, report issues, and seek support. Please let me know if this addresses the checklist item.
