### 2025-12-29 insysbio/LikelihoodProfiler.jl#70 @currocam
I cannot run the `joss.jl` code.

```julia
julia> using LikelihoodProfiler, OptimizationLBFGSB, OrdinaryDiffEqTsit5, PEtab, CICOBase

julia> using Plots

julia> model_name = "Boehm_JProteomeRes2014"
"Boehm_JProteomeRes2014"

julia> path_yaml = joinpath(@__DIR__, "../models/", "$model_name/$model_name.yaml")
"/Users/curro/Documents/Dev/LikelihoodProfiler.jl/joss/../models/Boehm_JProteomeRes2014/Boehm_JProteomeRes2014.yaml"

julia> petab_model = PEtabModel(path_yaml)
PEtabModel: Boehm_JProteomeRes2014 with 8 states and 10 parameters

julia> petab_problem = PEtabODEProblem(petab_model)

PEtabODEProblem: Boehm_JProteomeRes2014 with ODE-states 8 and 9 parameters to estimate
---------------- Problem options ---------------
Gradient method: ForwardDiff
Hessian method: ForwardDiff
ODE-solver nllh: Rodas5P
ODE-solver gradient: Rodas5P

julia> optprob = OptimizationProblem(petab_problem)
ERROR: No methods were found for the model function passed to the equation solver.
The function `f` needs to have dispatches, for example, for an ODEProblem
`f` must define either `f(u,p,t)` or `f(du,u,p,t)`. For more information
on how the model function `f` should be defined, consult the docstring for
the appropriate `AbstractSciMLFunction`.

Offending function: f
Stacktrace:
 [1] isinplace(f::PEtabODEProblem{…}, inplace_param_number::Int64, fname::String, iip_preferred::Bool; has_two_dispatches::Bool, isoptimization::Bool, outofplace_param_number::Int64)
   @ SciMLBase ~/.julia/packages/SciMLBase/pzpcE/src/utils.jl:254
 [2] isinplace (repeats 2 times)
   @ ~/.julia/packages/SciMLBase/pzpcE/src/utils.jl:246 [inlined]
 [3] #OptimizationFunction#183
   @ ~/.julia/packages/SciMLBase/pzpcE/src/scimlfunctions.jl:4278 [inlined]
 [4] OptimizationFunction
   @ ~/.julia/packages/SciMLBase/pzpcE/src/scimlfunctions.jl:4277 [inlined]
 [5] OptimizationProblem(::PEtabODEProblem{…}; kwargs::@Kwargs{})
   @ SciMLBase ~/.julia/packages/SciMLBase/pzpcE/src/problems/optimization_problems.jl:134
 [6] OptimizationProblem(::PEtabODEProblem{…})
   @ SciMLBase ~/.julia/packages/SciMLBase/pzpcE/src/problems/optimization_problems.jl:133
 [7] top-level scope
   @ REPL[7]:1
Some type information was truncated. Use `show(err)` to see complete types.
```

Is this some problem from my side?

Also, I can see that the last time the script `joss.jl` was updated (in the master branch) is after the last time the figures in the manuscript were updated.  Perhaps consider committing `Manifest.toml` and updated figures?
