### 2025-12-29 bcd9bd7
Optimization added to joss.jl
