+++ joss/Project.toml
@@ -1,6 +1,7 @@
 [deps]
 CICOBase = "33444335-3134-4342-4659-594331344435"
 LikelihoodProfiler = "93acb638-a083-5915-8dce-d129bc6a3f59"
+Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
 OptimizationLBFGSB = "22f7324a-a79d-40f2-bebe-3af60c77bd15"
 OrdinaryDiffEqTsit5 = "b1df2697-797e-41e3-8120-5422d3b24e4a"
 PEtab = "48d54b35-e43e-4a66-a5a1-dde6b987cf69"
