### 2025-12-30 d7fe7f3
CompatHelper: bump compat for PreallocationTools to 1, (keep existing compat)
