### 2026-01-04 #9501 @lockwo
Hi @dufourc1, @currocam, thanks for your reviewing work so far. I hope you had a good holiday. When you are back from the break, it would be great if you could share a brief update on the remaining things needed from the library for the review. I know the issues/reviewer requested changes for the paper were ongoing, so I just wanted to check in.
