### 2026-01-04 insysbio/LikelihoodProfiler.jl#70 @ivborissov
@currocam Thanks for checking — yes, this is intended. The example in `joss.jl` is fully deterministic. The initial parameter values come from the PEtab problem definition (optimal parameters), and the profiling methods used in the script (`IntegrationProfiler` and `CICOProfiler`) use them as starting values default. As a result, repeated runs should produce identical results.
