### 2026-01-05 insysbio/LikelihoodProfiler.jl#70 @currocam
Thanks! Feel free to close this issue
