### 2026-01-07 #9501 @lockwo
Great, thank you both for the quick response. @ivborissov although some GitHub issues are still open it looks like all the reviewers are in alignment (and the issues are basically complete, e.g. https://github.com/insysbio/LikelihoodProfiler.jl/issues/72#issuecomment-3671677584) and everyone has approved their checklists, is there further iteration you are working on for the paper that you would like to get done? (else with the 3 reviews complete, I can begin my steps before recommending acceptance).
