### 2026-01-07 #9501 @ivborissov
@lockwo Thank you very much. I’d like to thank again @salbalkus @currocam @dufourc1 for the careful review and constructive feedback — it significantly improved both the paper and the package. I don’t have any further updates planned for the manuscript or the code at this point and am happy for you to proceed with the next steps.
