### 2026-01-12 #9501 @lockwo
Hi @ivborissov, paper is looking good, few points:

- Just making sure any funding goes in an Acknowledgements section
- The axis names of the figures could be clarified (I assume they are the variable names/rows in a data frame)
- The length is ok (1,000 words is not a hard cutoff), just feel free to cut over add
