+++ joss/joss.jl
@@ -13,7 +13,7 @@ param_profile_prob = ProfileLikelihoodProblem(optprob, optpars; idxs=1:3)
 
 alg_integ = IntegrationProfiler(integrator = Tsit5(), integrator_opts = (dtmax=0.5, reltol=1e-3, abstol=1e-4),
   matrix_type = :identity, gamma=0., reoptimize=true, 
-  optimizer = LBFGSB(),optimizer_opts=(maxiters=10000,))
+  optimizer = LBFGSB(),optimizer_opts=(maxiters=11000,))
 
 ###################################### PARAMETER PROFILES ######################################
 
@@ -22,19 +22,19 @@ sol_param = solve(param_profile_prob, alg_integ, verbose=true)
 alg_cico = CICOProfiler(optimizer = :LN_NELDERMEAD, scan_tol = 1e-10)
 sol_cico = solve(param_profile_prob, alg_cico, verbose=true)
 
-w, h = 1000, 300
+w, h = 1100, 300
 
-p11 = plot(sol_param[1] , dpi=400, xguide="log10_Epo_degradation_BaF3", yguide="objective function", legend=false)
-p12 = plot(sol_param[2] , dpi=400, xguide="log10_k_exp_hetero", yguide="objective function", legend=false, title="Parameter Profiles")
-p13 = plot(sol_param[3] , dpi=400, xguide="log10_k_exp_homo", yguide="objective function", legend=:outerright)
+p11 = plot(sol_param[1] , dpi=400, xguide="log10(Epo_degradation_BaF3)", yguide="Objective function value", legend=false)
+p12 = plot(sol_param[2] , dpi=400, xguide="log10(k_exp_hetero)", yguide="Objective function value", legend=false, title="Parameter profiles (integration-based)")
+p13 = plot(sol_param[3] , dpi=400, xguide="log10(k_exp_homo)", yguide="Objective function value", legend=:outerright)
 
 p1 = plot(p11, p12, p13, layout=(1,3), size=(w,h), margins=5Plots.mm, dpi=400)
 
 savefig(p1, joinpath(@__DIR__, "param_profiles.png"))
 
-p21 = plot(sol_cico[1] , dpi=400, xguide="log10_Epo_degradation_BaF3", yguide="objective function", legend=false)
-p22 = plot(sol_cico[2] , dpi=400, xguide="log10_k_exp_hetero", yguide="objective function", legend=false, title="CICOProfiler (Confidence Intervals endpoints)")
-p23 = plot(sol_cico[3] , dpi=400, xguide="log10_k_exp_homo", yguide="objective function", legend=:outerright)
+p21 = plot(sol_cico[1] , dpi=400, xguide="log10(Epo_degradation_BaF3)", yguide="Objective function value", legend=false)
+p22 = plot(sol_cico[2] , dpi=400, xguide="log10(k_exp_hetero)", yguide="Objective function value", legend=false, title="Parameter profiles (CICO)")
+p23 = plot(sol_cico[3] , dpi=400, xguide="log10(k_exp_homo)", yguide="Objective function value", legend=:outerright)
 
 p2 = plot(p21, p22, p23, layout=(1,3), size=(w,h), margins=5Plots.mm, dpi=400)
 
@@ -56,11 +56,11 @@ func_profile_prob = ProfileLikelihoodProblem(optprob, optpars, [pSTAT5A_rel_optf
 
 sol_func = solve(func_profile_prob, alg_integ, verbose=true)
 
-w, h = 1000, 300
+w, h = 1100, 300
 
[…]
