### 2026-01-13 a8ee39e
joss: Acknowledgements added
