### 2026-01-14 #9501 @ivborissov
@lockwo Thanks for the guidance.

I’ve now:
- Created a tagged release of the software: **v1.3.1**
- Archived the reviewed version in Zenodo
- Verified that the Zenodo record has the correct title and author list (matching the manuscript), and added ORCID identifiers

The DOI for the archived version is:
**https://doi.org/10.5281/zenodo.18243712**

Please let me know if anything else is needed.
