### 2026-01-15 #9501 @lockwo
@ivborissov  - please let me know if you see any issues with the final proof that this generates
