### 2026-01-19 #9501 @crvernon
@ivborissov - I just need you to correct the following before I accept this for publication:

In your paper:

- [x] LINE 110:  The "r" (as in the R programming language) should be capitalized here.  You can do this in your bib file by wrapping the characters you wish to maintain formatting of in curly brackets.
- [x] LINE 122:  The "j" in "julia" should be capitalized.
- [x] LINE 127:  The "p" in "python" should be capitalized.

Let me know when these changes have been made.
