### 2026-01-19 7fc5fbd
JOSS: Julia, Python, R Capitalized
