+++ joss/paper.bib
@@ -46,7 +46,7 @@
 
 @article{Rackauckas2017,
   author = {C. Rackauckas and Q. Nie},
-  title = {DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
+  title = {DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in {Julia}},
   journal = {JORS},
   volume = {5},
   number = {1},
@@ -143,7 +143,7 @@
                   Dudkin, Erika and
                   Grein, Stephan and
                   Hasenauer, Jan},
-  title        = {pyPESTO - Parameter EStimation TOolbox for python},
+  title        = {pyPESTO - Parameter EStimation TOolbox for {Python}},
   month        = nov,
   year         = 2025,
   publisher    = {Zenodo},
@@ -159,11 +159,11 @@
 }
 
 @article{Kaschek2019,
- title={Dynamic Modeling, Parameter Estimation, and Uncertainty Analysis in R},
+ title={Dynamic Modeling, Parameter Estimation, and Uncertainty Analysis in {R}},
  volume={88},
  url={https://www.jstatsoft.org/index.php/jss/article/view/v088i10},
  doi={10.18637/jss.v088.i10},
- abstract={In a wide variety of research fields, dynamic modeling is employed as an instrument to learn and understand complex systems. The differential equations involved in this process are usually non-linear and depend on many parameters whose values determine the characteristics of the emergent system. The inverse problem, i.e., the inference or estimation of parameter values from observed data, is of interest from two points of view. First, the existence point of view, dealing with the question whether the system is able to reproduce the observed dynamics for any parameter values. Second, the identifiability point of view, investigating invariance of the prediction under change of parameter values, as well as the quantification of parameter uncertainty. In this paper, we present the R package dMod providing a framework for dealing with the inverse problem in dynamic systems modeled by ordinary differential equations. The uniqueness of the approach taken by dMod is to provide and propagate accurate derivatives computed from symbolic expressions wherever possible. This derivative information highly supports the convergence of optimization routines and enhances their numerical stability, a requirement for the applicability of sophisticated uncertainty analysis methods. Computational efficiency is achieved by automatic generation and execution of C code. The framework is object-oriented (S3) and provides a variety of functions to set up ordinary differential equation models, observation functions and parameter transformations for multi-conditional parameter estimation. The key elements of the framework and the methodology implemented in dMod are highlighted by an application on a three-compartment transporter model.},
[…]
