### 2026-01-19 #9501 @ivborissov
@crvernon Thanks for catching these. I’ve corrected the capitalization issues (R, Julia / Python in the references) and pushed the updated paper to the master branch.
