### 2026-01-20 #9501 @crvernon
🥳 Congratulations on your new publication @ivborissov ! Many thanks to @lockwo for editing and @salbalkus , @currocam , and @dufourc1 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
